feat: move proxy types
