+++ src/renderer/entities/proxy/index.ts
@@ -1,4 +1,3 @@
 export { proxyModel } from './model/proxy-model';
 export * from './lib/constants';
 export { ProxyAccount } from './ui/ProxyAccount/ProxyAccount';
-export { ProxyType } from './lib/types';
