+++ src/renderer/entities/proxy/lib/constants.ts
@@ -1,5 +1,4 @@
-import { ProxyType } from './types';
-import { AccountId, ProxyAccount } from '@shared/core';
+import { ProxyType } from '@shared/core';
 
 export const ProxyTypeName: Record<ProxyType, string> = {
   [ProxyType.ANY]: 'proxy.names.any',
@@ -11,16 +10,3 @@ export const ProxyTypeName: Record<ProxyType, string> = {
   [ProxyType.IDENTITY_JUDGEMENT]: 'proxy.names.identityJudgement',
   [ProxyType.NOMINATION_POOLS]: 'proxy.names.nominationPools',
 };
-
-export const ProxyTypeOperation: Record<ProxyType, string> = {
-  [ProxyType.ANY]: 'proxy.operations.any',
-  [ProxyType.NON_TRANSFER]: 'proxy.operations.nonTransfer',
-  [ProxyType.STAKING]: 'proxy.operations.staking',
-  [ProxyType.AUCTION]: 'proxy.operations.auction',
-  [ProxyType.CANCEL_PROXY]: 'proxy.operations.cancelProxy',
-  [ProxyType.GOVERNANCE]: 'proxy.operations.governance',
-  [ProxyType.IDENTITY_JUDGEMENT]: 'proxy.operations.identityJudgement',
-  [ProxyType.NOMINATION_POOLS]: 'proxy.operations.nominationPools',
-};
-
-export type ProxyStore = Record<AccountId, ProxyAccount[]>;
