+++ src/renderer/entities/proxy/lib/types.ts
@@ -1,10 +1,3 @@
-export const enum ProxyType {
-  ANY = 'Any',
-  NON_TRANSFER = 'NonTransfer',
-  STAKING = 'Staking',
-  AUCTION = 'Auction',
-  CANCEL_PROXY = 'CancelProxy',
-  GOVERNANCE = 'Governance',
-  IDENTITY_JUDGEMENT = 'IdentityJudgement',
-  NOMINATION_POOLS = 'NominationPools',
-}
+import type { AccountId, ProxyAccount } from '@shared/core';
+
+export type ProxyStore = Record<AccountId, ProxyAccount[]>;
