+++ src/renderer/features/notifications/NotificationProvider/ui/MultisigInvitedNotification.tsx
