+++ src/renderer/shared/core/index.ts
@@ -51,4 +51,5 @@ export type { Validator } from './types/validator';
 export { RewardsDestination } from './types/stake';
 export type { Stake, Unlocking } from './types/stake';
 
-export type { ProxyAccount, PartialProxyAccount, ProxyType } from './types/proxy';
+export { ProxyType } from './types/proxy';
+export type { ProxyAccount, PartialProxyAccount } from './types/proxy';
